requirements.txt
aioredis==1.3.1
alabaster==0.7.12
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.2
Automat==20.2.0
Babel==2.8.0
beautifulsoup4==4.9.0
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.9
daphne==2.4.1
diff-match-patch==20181111
Django==3.0.5
django-bootstrap-form==3.4
django-bootstrap4==1.1.1
django-filter==2.2.0
django-glrm==1.1.3
django-guardian==2.2.0
django-localflavor==3.0.1
django-phonenumber-field==4.0.0
django-reactor==1.4.0b0
django-seed==0.2.2
django-tables2==2.3.1
docutils==0.16
Faker==4.0.2
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
imagesize==1.2.0
incremental==17.5.0
Jinja2==2.11.1
MarkupSafe==1.1.1
msgpack==0.6.2
mysql-connector==2.2.9
mysqlclient==1.4.6
packaging==20.3
phonenumbers==8.12.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyHamcrest==2.0.2
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-stdnum==1.13
pytz==2019.3
requests==2.23.0
service-identity==18.1.0
six==1.14.0
snowballstemmer==2.0.0
soupsieve==2.0
Sphinx==3.0.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
text-unidecode==1.3
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.8
zope.interface==5.1.0
